backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.7
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
mercadopago==2.2.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pytest==7.4.3
python-dateutil==2.9.0.post0
python-jose==3.3.0
redis==5.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
vine==5.1.0
wcwidth==0.2.13
